[HomeScreen] 일기 수정 및 삭제 기능 구현

+++ feature/calendar/src/main/java/com/daylog/app/feature/calendar/CalendarViewModel.kt
@@ -61,9 +61,9 @@ class CalendarViewModel @Inject constructor(
                     .groupingBy { it.mood }
                     .eachCount(),
 
-                walkTotal = diaries.sumOf { it.walkCount },
+                walkTotal = diaries.sumOf { it.walkCount.toInt() },
 
-                snackTotal = diaries.sumOf { it.snackCount }
+                snackTotal = diaries.sumOf { it.snackCount.toInt()  }
             )
 
         }.stateIn(
